# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobExperienceLevel =
  (typeof JobExperienceLevel)[keyof typeof JobExperienceLevel];

export const JobExperienceLevel = {
  junior: "junior",
  mid: "mid",
  senior: "senior",
  lead: "lead",
  executive: "executive",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  description: string;
  requirements: string;
  skills: string[];
  experienceLevel: JobExperienceLevel;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status: JobStatus;
  matchCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateJobBodyExperienceLevel =
  (typeof CreateJobBodyExperienceLevel)[keyof typeof CreateJobBodyExperienceLevel];

export const CreateJobBodyExperienceLevel = {
  junior: "junior",
  mid: "mid",
  senior: "senior",
  lead: "lead",
  executive: "executive",
} as const;

export type CreateJobBodyStatus =
  (typeof CreateJobBodyStatus)[keyof typeof CreateJobBodyStatus];

export const CreateJobBodyStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface CreateJobBody {
  title: string;
  company: string;
  location: string;
  description: string;
  requirements: string;
  skills: string[];
  experienceLevel: CreateJobBodyExperienceLevel;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status?: CreateJobBodyStatus;
}

export type UpdateJobBodyExperienceLevel =
  (typeof UpdateJobBodyExperienceLevel)[keyof typeof UpdateJobBodyExperienceLevel];

export const UpdateJobBodyExperienceLevel = {
  junior: "junior",
  mid: "mid",
  senior: "senior",
  lead: "lead",
  executive: "executive",
} as const;

export type UpdateJobBodyStatus =
  (typeof UpdateJobBodyStatus)[keyof typeof UpdateJobBodyStatus];

export const UpdateJobBodyStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface UpdateJobBody {
  title?: string;
  company?: string;
  location?: string;
  description?: string;
  requirements?: string;
  skills?: string[];
  experienceLevel?: UpdateJobBodyExperienceLevel;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status?: UpdateJobBodyStatus;
}

export type CandidateStatus =
  (typeof CandidateStatus)[keyof typeof CandidateStatus];

export const CandidateStatus = {
  active: "active",
  inactive: "inactive",
  hired: "hired",
} as const;

export interface Candidate {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  currentTitle: string;
  summary: string;
  skills: string[];
  experienceYears: number;
  education: string;
  location: string;
  status: CandidateStatus;
  matchCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateCandidateBodyStatus =
  (typeof CreateCandidateBodyStatus)[keyof typeof CreateCandidateBodyStatus];

export const CreateCandidateBodyStatus = {
  active: "active",
  inactive: "inactive",
  hired: "hired",
} as const;

export interface CreateCandidateBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  currentTitle: string;
  summary: string;
  skills: string[];
  experienceYears: number;
  education: string;
  location: string;
  status?: CreateCandidateBodyStatus;
}

export type UpdateCandidateBodyStatus =
  (typeof UpdateCandidateBodyStatus)[keyof typeof UpdateCandidateBodyStatus];

export const UpdateCandidateBodyStatus = {
  active: "active",
  inactive: "inactive",
  hired: "hired",
} as const;

export interface UpdateCandidateBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  currentTitle?: string;
  summary?: string;
  skills?: string[];
  experienceYears?: number;
  education?: string;
  location?: string;
  status?: UpdateCandidateBodyStatus;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  pending: "pending",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface Match {
  id: number;
  jobId: number;
  candidateId: number;
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  assessment: string;
  matchedSkills: string[];
  missingSkills: string[];
  status: MatchStatus;
  createdAt: string;
}

export type MatchWithCandidateStatus =
  (typeof MatchWithCandidateStatus)[keyof typeof MatchWithCandidateStatus];

export const MatchWithCandidateStatus = {
  pending: "pending",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface MatchWithCandidate {
  id: number;
  jobId: number;
  candidateId: number;
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  assessment: string;
  matchedSkills: string[];
  missingSkills: string[];
  status: MatchWithCandidateStatus;
  createdAt: string;
  candidateName: string;
  candidateTitle: string;
  candidateEmail: string;
}

export type MatchWithJobStatus =
  (typeof MatchWithJobStatus)[keyof typeof MatchWithJobStatus];

export const MatchWithJobStatus = {
  pending: "pending",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface MatchWithJob {
  id: number;
  jobId: number;
  candidateId: number;
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  assessment: string;
  matchedSkills: string[];
  missingSkills: string[];
  status: MatchWithJobStatus;
  createdAt: string;
  jobTitle: string;
  jobCompany: string;
}

export type UpdateMatchStatusBodyStatus =
  (typeof UpdateMatchStatusBodyStatus)[keyof typeof UpdateMatchStatusBodyStatus];

export const UpdateMatchStatusBodyStatus = {
  pending: "pending",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface UpdateMatchStatusBody {
  status: UpdateMatchStatusBodyStatus;
}

export interface DashboardStats {
  totalJobs: number;
  openJobs: number;
  totalCandidates: number;
  activeCandidates: number;
  totalMatches: number;
  avgMatchScore: number;
  shortlistedCount: number;
  hiredCount: number;
}

export interface RecentMatch {
  id: number;
  jobTitle: string;
  candidateName: string;
  overallScore: number;
  status: string;
  createdAt: string;
}

export interface TopCandidate {
  candidateId: number;
  candidateName: string;
  candidateTitle: string;
  avgScore: number;
  matchCount: number;
  topSkills: string[];
}

export interface SkillDemand {
  skill: string;
  jobCount: number;
  candidateCount: number;
}

export type ListJobsParams = {
  status?: ListJobsStatus;
  search?: string;
};

export type ListJobsStatus =
  (typeof ListJobsStatus)[keyof typeof ListJobsStatus];

export const ListJobsStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export type ListCandidatesParams = {
  status?: ListCandidatesStatus;
  search?: string;
};

export type ListCandidatesStatus =
  (typeof ListCandidatesStatus)[keyof typeof ListCandidatesStatus];

export const ListCandidatesStatus = {
  active: "active",
  inactive: "inactive",
  hired: "hired",
} as const;

export type GetRecentMatchesParams = {
  limit?: number;
};

export type GetTopCandidatesParams = {
  limit?: number;
};