# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job matching API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, draft]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, hired]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateBody"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateBody"
      responses:
        "200":
          description: Candidate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        description:
          type: string
        requirements:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        status:
          type: string
          enum: [open, closed, draft]
        matchCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - company
        - location
        - description
        - requirements
        - skills
        - experienceLevel
        - status
        - matchCount
        - createdAt
        - updatedAt
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        description:
          type: string
        requirements:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        status:
          type: string
          enum: [open, closed, draft]
      required:
        - title
        - company
        - location
        - description
        - requirements
        - skills
        - experienceLevel
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        description:
          type: string
        requirements:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        status:
          type: string
          enum: [open, closed, draft]
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentTitle:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        education:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [active, inactive, hired]
        matchCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - currentTitle
        - summary
        - skills
        - experienceYears
        - education
        - location
        - status
        - matchCount
        - createdAt
        - updatedAt
    CreateCandidateBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentTitle:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        education:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [active, inactive, hired]
      required:
        - name
        - email
        - currentTitle
        - summary
        - skills
        - experienceYears
        - education
        - location
    UpdateCandidateBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentTitle:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        education:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [active, inactive, hired]
...