# lib/api-zod/src/generated/types/createCandidateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCandidateBodyStatus } from "./createCandidateBodyStatus";

export interface CreateCandidateBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  currentTitle: string;
  summary: string;
  skills: string[];
  experienceYears: number;
  education: string;
  location: string;
  status?: CreateCandidateBodyStatus;
}