# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateJobBodyExperienceLevel } from "./updateJobBodyExperienceLevel";
import type { UpdateJobBodyStatus } from "./updateJobBodyStatus";

export interface UpdateJobBody {
  title?: string;
  company?: string;
  location?: string;
  description?: string;
  requirements?: string;
  skills?: string[];
  experienceLevel?: UpdateJobBodyExperienceLevel;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status?: UpdateJobBodyStatus;
}